==> code/notebooks/1-DatasetReport.ipynb <==
Author: Kevin ALBERT

Created: Nov 2020

import pandas as pd
import dtale
import pandas_profiling as pp
from IPython.display import Javascript

# check versions
!conda -V
!python -V
!conda list |grep pandas
!conda list |grep pandas_profiling
!conda list |grep dtale

### load data

# https://biolib.com/shd2020/Diabetes-Bioinformatics-Data/
# manually copied files into these folders
synthetic_df = pd.read_csv("../../data/bronze/synthetic_data.csv")
features_description = pd.read_csv("../../data/bronze/feature_descriptions.csv")

### summary

# concise summary (shape, memory use, data types, nan's)
synthetic_df.info()

### static report

reportFile = "../../data/report/synthetic_diabetes_report.html"

%%time
# full report on 100% records (no correlation matrix stuff)
pp.ProfileReport(df=synthetic_df.sample(frac=1),
                 minimal=False,
                 progress_bar=False,
                 correlations={"cramers": {"calculate": False}}).to_file(reportFile)

# open the report (*.html)
display(Javascript('window.open("{url}");'.format(url=reportFile)))

### interactive report

# start webapp (change IP, port)
d = dtale.show(synthetic_df, host="13.93.37.217", port="40000", ignore_duplicate=True, drop_index=True, reaper_on=False)

# show all running instances
d.main_url()

# stop webapp
# d.kill()

### data checks

# count distinct (=unique) observations (+ missing), sorted (high cardinality > 390)
synthetic_df.nunique(dropna=False).sort_values(ascending=False)
