==> code/notebooks/3-DataGraph.ipynb <==
Author: Kevin ALBERT

Created: Nov 2020

from py2neo import Graph, Node, Relationship
import pandas as pd
from IPython.display import Javascript
# from fuzzywuzzy import process, fuzz

import platform
import psutil
import os

conda_version = ! conda -V
print(f"conda : {conda_version[0].split()[1]}")
pip_version = ! pip -V
print(f"pip   : {pip_version[0].split()[1]}")
python_version = ! python -V
print(f"python: {python_version[0].split()[1]}")
pandas_version = ! pip list |grep -i pandas
print(f"pandas: {pandas_version[0].split()[1]}")
py2neo_version = ! pip list |grep -i py2neo
print(f"py2neo: {py2neo_version[0].split()[1]}")

server  = "13.93.37.217"
port    = "7687"
user    = "neo4j"
passw   = "digityser"
db_name = "neo4j"        # default name (v4.x)

# load graph connection instance
graph = Graph(host=server, auth=(user, passw), name=db_name, encrypted=False)

### prepare data

# import the dataset
synthetic_df = pd.read_parquet("../../data/silver/synthetic_data_processed.parquet")

# creating unique patient ID
synthetic_df["id"] = synthetic_df.index
# real_df["id"] = real_df.index

meds_list = ['metformin', 'repaglinide', 'nateglinide',
       'chlorpropamide', 'glimepiride', 'acetohexamide', 'glipizide',
       'glyburide', 'tolbutamide', 'pioglitazone', 'rosiglitazone', 'acarbose',
       'miglitol', 'troglitazone', 'tolazamide', 'examide', 'citoglipton',
       'insulin', 'glyburide_metformin', 'glipizide_metformin',
       'glimepiride_pioglitazone', 'metformin_rosiglitazone',
       'metformin_pioglitazone']
for meds in meds_list:
    synthetic_df[meds+"_medicine"] = synthetic_df[meds].apply(lambda x: 'None' if (x=='No') else meds)

synthetic_df['readmitted'].apply(lambda x: 'None' if (x=='No') else meds)

# we make severe, normal diabetes severity levels
hba1c = {'None':'None', 'Norm':'Normal', '>8':'Severe_Diabetes', '>7':'Diabetes'}
synthetic_df["A1Cresult"] = synthetic_df["A1Cresult"].map(hba1c)

# save file to /import
file = "synthetic_data_processed.csv"
synthetic_df.to_csv("../../neo4j/import/"+file, sep=',', index=False)

### import data in DB

graph.run("CREATE CONSTRAINT ON (p:Patient) ASSERT p.id IS UNIQUE")

graph.run("CREATE CONSTRAINT ON (r:Race) ASSERT r.value IS UNIQUE")

query = """
USING PERIODIC COMMIT 1000
LOAD CSV WITH HEADERS FROM 'file:///"""+file+"""' AS line FIELDTERMINATOR ','
MERGE (p:Patient {id:(line.id)})
MERGE (r:Race {value:(line.race)})
MERGE (p)-[:HAS_RACE]->(r)
"""
display(graph.run(query).stats())

graph.run("CREATE CONSTRAINT ON (g:Gender) ASSERT g.value IS UNIQUE")

query = """
USING PERIODIC COMMIT 1000
LOAD CSV WITH HEADERS FROM 'file:///"""+file+"""' AS line FIELDTERMINATOR ','
MATCH (p:Patient {id:(line.id)})
MERGE (g:Gender {value:(line.gender)})
MERGE (p)-[:HAS_GENDER]->(g)
"""
display(graph.run(query).stats())

graph.run("CREATE CONSTRAINT ON (ag:AgeGroup) ASSERT ag.value IS UNIQUE")

query = """
USING PERIODIC COMMIT 1000
LOAD CSV WITH HEADERS FROM 'file:///"""+file+"""' AS line FIELDTERMINATOR ','
MATCH (p:Patient {id:(line.id)})
MERGE (ag:AgeGroup {value:(line.age)})
MERGE (p)-[:HAS_AGE]->(ag)
"""
display(graph.run(query).stats())

# graph.run("CREATE CONSTRAINT ON (m:Medicine) ASSERT m.value IS UNIQUE")
graph.run("CREATE INDEX ON :Medicine(value)")

for meds in meds_list:
    print(meds)
    query = """
    USING PERIODIC COMMIT 1000
    LOAD CSV WITH HEADERS FROM 'file:///"""+file+"""' AS line FIELDTERMINATOR ','
    WITH line.id AS id, line."""+meds+"""_medicine AS medicine, line."""+meds+""" AS drugdose
    WHERE medicine <> "None" AND drugdose <> "No"
    MATCH (p:Patient {id:(id)})
    MERGE (m:Medicine {value:(medicine), drugdose:(drugdose)})
    MERGE (p)-[:USE_MEDICINE]->(m)
    """
    display(graph.run(query).stats())

# graph.run("CREATE CONSTRAINT ON (hd:HospitalDays) ASSERT hd.severity IS UNIQUE")
graph.run("CREATE INDEX ON :HospitalDays(severity)")

query = """
USING PERIODIC COMMIT 1000
LOAD CSV WITH HEADERS FROM 'file:///"""+file+"""' AS line FIELDTERMINATOR ','
MATCH (p:Patient {id:(line.id)})
MERGE (hd:HospitalDays {severity:(line.time_in_hospital_severitylvl), value:(line.time_in_hospital)})
MERGE (p)-[:HOSPITAL_STAY]->(hd)
"""
display(graph.run(query).stats())

graph.run("CREATE CONSTRAINT ON (h:HbA1c) ASSERT h.value IS UNIQUE")

query = """
USING PERIODIC COMMIT 1000
LOAD CSV WITH HEADERS FROM 'file:///"""+file+"""' AS line FIELDTERMINATOR ','
WITH line
WHERE line.A1Cresult <> "None"
MATCH (p:Patient {id:(line.id)})
MERGE (h:HbA1c {value:(line.A1Cresult)})
MERGE (p)-[:HBA1C_STATUS]->(h)
"""
display(graph.run(query).stats())

graph.run("CREATE CONSTRAINT ON (ra:Readmitted) ASSERT ra.value IS UNIQUE")

query = """
USING PERIODIC COMMIT 1000
LOAD CSV WITH HEADERS FROM 'file:///"""+file+"""' AS line FIELDTERMINATOR ','
MATCH (p:Patient {id:(line.id)})
MERGE (ra:Readmitted {value:(line.readmitted)})
MERGE (p)-[:RECURRENT]->(ra)
"""
display(graph.run(query).stats())

## Querying

%%time
query = """
MATCH (n:HbA1c)<-[:HBA1C_STATUS]-(p:Patient)-[:USE_MEDICINE]->(m:Medicine)
WHERE m.value='insulin' AND m.drugdose='Down'
WITH m, p
MATCH (p)-[:HAS_AGE]->(ag:AgeGroup)
WHERE ag.value='80-90'
WITH p, ag
MATCH (n:HbA1c)<-[:HBA1C_STATUS]-(p)-[:USE_MEDICINE]->(meds:Medicine)
MATCH (p)-[:RECURRENT]->(ra:Readmitted)
//WHERE meds.drugdose='Down' OR meds.drugdose='Steady' 
RETURN ag.value, meds, n.value, COLLECT(DISTINCT p.id), ra.value ORDER BY ag.value DESC LIMIT 200
"""
graph.run(query).to_data_frame()

## Sunday morning breakthrough :-)

#### readmitted = 'NO'

%%time
query = """
MATCH (n:HbA1c)<-[:HBA1C_STATUS]-(p:Patient)-[:USE_MEDICINE]->(m:Medicine)
MATCH (p)-[:RECURRENT]->(ra:Readmitted {value:'NO'})
WHERE m.drugdose='Down' OR m.drugdose='Steady' OR m.drugedose='Up'
RETURN p.id AS patient, m.value as medicine, n.value AS Hemoglobine
"""
graph.run(query).to_data_frame()

#### readmitted = '<30'

%%time
query = """
MATCH (n:HbA1c)<-[:HBA1C_STATUS]-(p:Patient)-[:USE_MEDICINE]->(m:Medicine)
MATCH (p)-[:RECURRENT]->(ra:Readmitted {value:'<30'})
WHERE m.drugdose='Down' OR m.drugdose='Steady' OR m.drugedose='Up'
RETURN p.id AS patient, m.value as medicine, n.value AS Hemoglobine
"""
graph.run(query).to_data_frame()

%%time
query = """
MATCH (n:HbA1c)<-[:HBA1C_STATUS]-(p:Patient)-[:USE_MEDICINE]->(m:Medicine)
MATCH (p)-[:RECURRENT]->(ra:Readmitted {value:'>30'})
WHERE m.drugdose='Down' OR m.drugdose='Steady' OR m.drugedose='Up'
RETURN p.id AS patient, m.value as medicine, n.value AS Hemoglobine
"""
graph.run(query).to_data_frame()

### Can you explore which treatments are effective, and which ones are not?

### What are the most important factors for hospital readmission?

### Is the best cure insulin or a combination of drugs?

### What's the best treatment for a newly diagnosed patient based on patient's features and the response to certain drugs?

### reset DB

# delete database and restart (wait ~2min)
import os
os.system(" cd ../.. && \
            sudo docker-compose down && \
            sudo rm -Rf neo4j/data/databases/neo4j && \
            sudo rm -Rf neo4j/data/transactions/neo4j && \
            sudo docker-compose up --build &")
