==> code/notebooks/5-MachineLearning-experiments-change.ipynb <==
Author: Kevin ALBERT

Created: Nov 2020

# import logging
import os
import logging
import pandas as pd
import numpy as np
import json
import requests
import joblib

from azureml.core import Workspace, Dataset, Datastore, Run
from azureml.core.experiment import Experiment
from azureml.data.datapath import DataPath
from azureml.core.compute import ComputeTarget, AmlCompute, AksCompute
from azureml.core.model import Model, InferenceConfig
from azureml.train.automl import AutoMLConfig
from azureml.train.automl.run import AutoMLRun
from azureml.widgets import RunDetails
from azureml.core.webservice import Webservice, AciWebservice, AksWebservice
from azureml.exceptions import WebserviceException
from azureml.core.environment import Environment
from azureml.train.estimator import Estimator
from azureml.core.conda_dependencies import CondaDependencies
from azureml.train.hyperdrive.runconfig import HyperDriveConfig
from azureml.train.hyperdrive.sampling import RandomParameterSampling, GridParameterSampling
from azureml.train.hyperdrive.run import PrimaryMetricGoal
from azureml.train.hyperdrive.parameter_expressions import choice
from azureml.core.runconfig import RunConfiguration
from azureml.pipeline.core import PipelineData, Pipeline
from azureml.pipeline.steps import PythonScriptStep, EstimatorStep
from azureml.pipeline.core.run import PipelineRun
from azureml.core.authentication import InteractiveLoginAuthentication
from azureml.explain.model._internal.explanation_client import ExplanationClient
import azureml.core
print("azureml.core version:", azureml.core.__version__)

# pd.describe_option('display')            # show all pandas options, parameters can slow down notebook
pd.set_option('display.max_colwidth', 100) # default 50, the maximum width in characters of a column
pd.set_option('display.max_columns', 40)   # default 20, the maximum amount of columns in view 
pd.set_option('display.max_rows', 60)      # default 60, the maximum amount of rows in view

!pip list |grep -i azureml

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
%matplotlib inline
import seaborn as sns
from sklearn.model_selection import train_test_split
from sklearn.model_selection import cross_val_score
from sklearn.preprocessing import StandardScaler
from sklearn.tree import DecisionTreeRegressor
from sklearn.ensemble import RandomForestRegressor
from sklearn.dummy import DummyRegressor
from sklearn.linear_model import LinearRegression
from sklearn.model_selection import KFold
from sklearn.model_selection import StratifiedKFold
from sklearn.model_selection import ShuffleSplit
from sklearn.model_selection import GridSearchCV
from sklearn.metrics import mean_squared_error
from sklearn.metrics import mean_absolute_error
from sklearn.metrics import r2_score
#import eli5
#from eli5.sklearn import PermutationImportance
# import scikitplot as skplt
from sklearn.decomposition import PCA

sns.set_style('whitegrid')
pd.set_option('display.max_columns', None) # display all columns

import warnings
warnings.filterwarnings('ignore')

# load the workspace
ws = Workspace.from_config()

# choose an experiment name
experiment = Experiment(ws, 'automl-classification-synthetic-health-data')

ds = Datastore.register_azure_blob_container(
    workspace=ws,
    datastore_name="datalakestoragegen2",
    container_name="datalake",
    account_name="datalake27112020",
    account_key="WJ4lTl5w9ze3hberxHsnNoWON5DTZJyajggoLY3j7WgsDDFm5w/NPuDAfO4Po/bNellztxilXm2Gpo9GzEzxdA==",
    create_if_not_exists=False)
# list available datastores
ws.datastores

# setup parquet file(s) into a tabular dataset
ds_path = [DataPath(ds, 'silver/synthetic_data_processed_remove_hba1c.parquet')] # {path/*.parquet}
dataset = Dataset.Tabular.from_parquet_files(path=ds_path)
# show dataset settings
dataset

# dataset.to_pandas_dataframe()["A1Cresult"]

automl_settings = {
    "enable_early_stopping":True,
    "experiment_timeout_hours":0.5,
    "iterations":10, # number of runs ex: 20
    "iteration_timeout_minutes":5,
    "max_concurrent_iterations":1,
    "max_cores_per_iteration":-1,
    #"experiment_exit_score":0.9920,
    "model_explainability":True,
    "n_cross_validations":5, # make 10 if small dataset, else 5
    "primary_metric":'AUC_weighted',
    "featurization":'auto',
    "verbosity":logging.INFO, # {INFO, DEBUG, CRITICAL, ERROR, WARNING} -- debug_log=<*.log>
    "enable_dnn":False,
}

automl_config = AutoMLConfig(task='classification',
                             debug_log='automl_errors.log',
                             compute_target='local', # {training_cluster or 'local'}
                             #blacklist_models=['KNN','LinearSVM'],
#                              environment_definition=training_env,
                             enable_onnx_compatible_models=True,
                             training_data=dataset,
                             label_column_name="readmitted",
#                              training_data=training_data,
#                              validation_data=validation_data,
#                              label_column_name="readmitted",
                             **automl_settings
                            )
# ouputs "model.pkl" and "automl_errors.log"

automl_run = experiment.submit(automl_config, show_output=True)

### optional: retrieve specific run

# runId = 'AutoML_891419fd-d69c-4a91-b536-f008adcb800c'
# automl_run = AutoMLRun(experiment, run_id=runId)

### results: explore the best pipeline

RunDetails(automl_run).show()
automl_run.wait_for_completion() # get more parameter info

best_run, fitted_model = automl_run.get_output()
# pipeline steps
for step in fitted_model.named_steps:
    print(step)

# model properties
fitted_model.named_steps

# show all metrics
best_run.get_metrics()

# best_run.download_file('residuals')

# best_run.download_file('predicted_true')

best_run.get_details()

client = ExplanationClient.from_run(best_run)
engineered_explanations = client.download_model_explanation(raw=False)
feature_importance = engineered_explanations.get_feature_importance_dict() # get model feature importance values
feature_importance

columns = ["modelFeatureImportance_name", "modelFeatureImportance_value"]
fi = pd.DataFrame(list(feature_importance.items()), columns=columns) # transform dict into DataFrame
fi

fi = fi[fi.iloc[:,1] >= 0.01] # remove small or zero values
fi

fi_sum = fi.iloc[:,1].sum()
relativeWeight = lambda x: x/fi_sum # normalize percentage values of weights
fi["modelFeatureImportance_relativeWeight"] = fi.iloc[:,1].apply(relativeWeight)
fi

# second run on >30

# setup parquet file(s) into a tabular dataset
ds_path = [DataPath(ds, 'silver/synthetic_data_processed_morethan30.parquet')] # {path/*.parquet}
dataset = Dataset.Tabular.from_parquet_files(path=ds_path)
# show dataset settings
dataset

# choose an experiment name
experiment = Experiment(ws, 'automl-classification-synthetic-change')

automl_settings = {
    "enable_early_stopping":True,
    "experiment_timeout_hours":0.5,
    "iterations":10, # number of runs ex: 20
    "iteration_timeout_minutes":5,
    "max_concurrent_iterations":1,
    "max_cores_per_iteration":-1,
    #"experiment_exit_score":0.9920,
    "model_explainability":True,
    "n_cross_validations":5, # make 10 if small dataset, else 5
    "primary_metric":'AUC_weighted',
    "featurization":'auto',
    "verbosity":logging.INFO, # {INFO, DEBUG, CRITICAL, ERROR, WARNING} -- debug_log=<*.log>
    "enable_dnn":False,
}

automl_config = AutoMLConfig(task='classification',
                             debug_log='automl_errors.log',
                             compute_target='local', # {training_cluster or 'local'}
                             #blacklist_models=['KNN','LinearSVM'],
#                              environment_definition=training_env,
                             enable_onnx_compatible_models=True,
                             training_data=dataset,
                             label_column_name="change",
#                              training_data=training_data,
#                              validation_data=validation_data,
#                              label_column_name="readmitted",
                             **automl_settings
                            )
# ouputs "model.pkl" and "automl_errors.log"

automl_run = experiment.submit(automl_config, show_output=True)

RunDetails(automl_run).show()
automl_run.wait_for_completion() # get more parameter info

best_run, fitted_model = automl_run.get_output()

# pipeline steps
for step in fitted_model.named_steps:
    print(step)

# model properties
fitted_model.named_steps

# show all metrics
best_run.get_metrics()

best_run.get_details()

client = ExplanationClient.from_run(best_run)
engineered_explanations = client.download_model_explanation(raw=False)
feature_importance = engineered_explanations.get_feature_importance_dict() # get model feature importance values
feature_importance

columns = ["modelFeatureImportance_name", "modelFeatureImportance_value"]
fi = pd.DataFrame(list(feature_importance.items()), columns=columns) # transform dict into DataFrame
fi

fi = fi[fi.iloc[:,1] >= 0.01] # remove small or zero values
fi

fi_sum = fi.iloc[:,1].sum()
relativeWeight = lambda x: x/fi_sum # normalize percentage values of weights
fi["modelFeatureImportance_relativeWeight"] = fi.iloc[:,1].apply(relativeWeight)
fi

# third run on No

# setup parquet file(s) into a tabular dataset
ds_path = [DataPath(ds, 'silver/synthetic_data_processed_withNo.parquet')] # {path/*.parquet}
dataset = Dataset.Tabular.from_parquet_files(path=ds_path)
# show dataset settings
dataset

dataset.to_pandas_dataframe().shape

# choose an experiment name
experiment = Experiment(ws, 'automl-classification-synthetic-change')

automl_settings = {
    "enable_early_stopping":True,
    "experiment_timeout_hours":0.5,
    "iterations":10, # number of runs ex: 20
    "iteration_timeout_minutes":5,
    "max_concurrent_iterations":1,
    "max_cores_per_iteration":-1,
    #"experiment_exit_score":0.9920,
    "model_explainability":True,
    "n_cross_validations":5, # make 10 if small dataset, else 5
    "primary_metric":'AUC_weighted',
    "featurization":'auto',
    "verbosity":logging.INFO, # {INFO, DEBUG, CRITICAL, ERROR, WARNING} -- debug_log=<*.log>
    "enable_dnn":False,
}

automl_config = AutoMLConfig(task='classification',
                             debug_log='automl_errors.log',
                             compute_target='local', # {training_cluster or 'local'}
                             #blacklist_models=['KNN','LinearSVM'],
#                              environment_definition=training_env,
                             enable_onnx_compatible_models=True,
                             training_data=dataset,
                             label_column_name="change",
#                              training_data=training_data,
#                              validation_data=validation_data,
#                              label_column_name="readmitted",
                             **automl_settings
                            )
# ouputs "model.pkl" and "automl_errors.log"

automl_run = experiment.submit(automl_config, show_output=True)

RunDetails(automl_run).show()
automl_run.wait_for_completion() # get more parameter info

best_run, fitted_model = automl_run.get_output()

# pipeline steps
for step in fitted_model.named_steps:
    print(step)

# model properties
fitted_model.named_steps

# show all metrics
best_run.get_metrics()

best_run.get_details()

client = ExplanationClient.from_run(best_run)
engineered_explanations = client.download_model_explanation(raw=False)
feature_importance = engineered_explanations.get_feature_importance_dict() # get model feature importance values
feature_importance

columns = ["modelFeatureImportance_name", "modelFeatureImportance_value"]
fi = pd.DataFrame(list(feature_importance.items()), columns=columns) # transform dict into DataFrame
fi

fi = fi[fi.iloc[:,1] >= 0.01] # remove small or zero values
fi

fi_sum = fi.iloc[:,1].sum()
relativeWeight = lambda x: x/fi_sum # normalize percentage values of weights
fi["modelFeatureImportance_relativeWeight"] = fi.iloc[:,1].apply(relativeWeight)
fi
